+++ src/test/java/com/neo4jParquet/Neo4jParquetTest.java
@@ -4,12 +4,6 @@ import org.assertj.core.api.Assertions;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 
-import java.io.File;
-
-import org.junit.jupiter.api.AfterAll;
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeAll;
-import org.junit.Rule;
 import org.junit.Test;
 import org.neo4j.driver.Driver;
 import org.neo4j.driver.GraphDatabase;
@@ -43,6 +37,12 @@ public class Neo4jParquetTest {
 
             result=session.run("call com.neo4jparquet.readParquet(\"src/main/resources/example.parquet\") yield value with value where value.Centroid is not null return value.Centroid as centroid limit 1");
             assertEquals("\"34.75298985475595 -6.257732428506092\"", result.single().get("centroid").toString());
+
+            result=session.run("call com.neo4jparquet.readParquetWKT(\"src/main/resources/examples.parquet\") yield value return value.errorMessage as errorMessage limit 1");
+            assertEquals("\"-1\"", result.single().get("errorMessage").toString());
+
+            result=session.run("call com.neo4jparquet.readParquetWKT(\"src/main/resources/multilinestring.parquet\") yield value return value.Geometry as Geometry limit 1");
+            assertEquals("\"MULTILINESTRING(( 100.0 0.0), ( 101.0 1.0), ( 102.0 2.0), ( 103.0 3.0))\"", result.single().get("Geometry").toString());
 
         }
    }