+++ CHANGELOG.md
@@ -1,2 +1,6 @@
+# CHANGELOG - 0.20.0 - 2023-12-28
+* [ Added ] Added Support for MULTILINESTRING
+* [Changed] Updated Error Messages in Code and Documentation.md file. Error returns a json key-value pair with the "errorMessage" as the key.
+
 # CHANGELOG - 0.10.0 - 2023-12-28
 * [ Added ] Initial release for experimentation.